add extra class
