+++ napari/_qt/layers/qt_image_layer.py
@@ -1,6 +1,7 @@
 from qtpy.QtWidgets import QHBoxLayout
 from .. import QVRangeSlider
-from .qt_base_layer import QtLayerControls, QtLayerProperties
+from .qt_base_layer import QtLayerControls
+from .qt_image_base_layer import QtImageBaseProperties
 from qtpy.QtCore import Qt
 from qtpy.QtWidgets import QLabel, QComboBox
 from ...layers.image._constants import Interpolation, Rendering
@@ -52,78 +53,45 @@ class QtImageControls(QtLayerControls):
         self.layer.status = self.layer._contrast_limits_msg
 
 
-class QtImageProperties(QtLayerProperties):
+class QtImageProperties(QtImageBaseProperties):
     def __init__(self, layer):
         super().__init__(layer)
 
-        if hasattr(self.layer, 'colormap'):
-            self.layer.events.colormap.connect(self._on_colormap_change)
-            row = self.grid_layout.rowCount()
-            comboBox = QComboBox()
-            for cmap in self.layer.colormaps:
-                comboBox.addItem(cmap)
-            comboBox._allitems = set(self.layer.colormaps)
-            index = comboBox.findText(
-                self.layer.colormap[0], Qt.MatchFixedString
-            )
-            comboBox.setCurrentIndex(index)
-            comboBox.activated[str].connect(
-                lambda text=comboBox: self.changeColor(text)
-            )
-            self.grid_layout.addWidget(
-                QLabel('colormap:'), row, self.name_column
-            )
-            self.grid_layout.addWidget(comboBox, row, self.property_column)
-            self.colormap_combobox = comboBox
+        self.layer.events.interpolation.connect(self._on_interpolation_change)
+        row = self.grid_layout.rowCount()
+        interp_comboBox = QComboBox()
+        for interp in Interpolation:
+            interp_comboBox.addItem(str(interp))
+        index = interp_comboBox.findText(
+            self.layer.interpolation, Qt.MatchFixedString
+        )
+        interp_comboBox.setCurrentIndex(index)
+        interp_comboBox.activated[str].connect(
+            lambda text=interp_comboBox: self.changeInterpolation(text)
+        )
+        self.interpComboBox = interp_comboBox
+        self.grid_layout.addWidget(
+            QLabel('interpolation:'), row, self.name_column
+        )
+        self.grid_layout.addWidget(interp_comboBox, row, self.property_column)
 
-        if hasattr(self.layer, 'interpolation'):
-            self.layer.events.interpolation.connect(
-                self._on_interpolation_change
-            )
-            row = self.grid_layout.rowCount()
-            interp_comboBox = QComboBox()
-            for interp in Interpolation:
-                interp_comboBox.addItem(str(interp))
-            index = interp_comboBox.findText(
-                self.layer.interpolation, Qt.MatchFixedString
-            )
-            interp_comboBox.setCurrentIndex(index)
-            interp_comboBox.activated[str].connect(
-                lambda text=interp_comboBox: self.changeInterpolation(text)
-            )
-            self.interpComboBox = interp_comboBox
-            self.grid_layout.addWidget(
-                QLabel('interpolation:'), row, self.name_column
-            )
-            self.grid_layout.addWidget(
-                interp_comboBox, row, self.property_column
-            )
-
-        if hasattr(self.layer, 'rendering'):
-            self.layer.events.rendering.connect(self._on_rendering_change)
-            row = self.grid_layout.rowCount()
-            renderComboBox = QComboBox()
-            for render in Rendering:
-                renderComboBox.addItem(str(render))
-            index = renderComboBox.findText(
-                self.layer.rendering, Qt.MatchFixedString
-            )
-            renderComboBox.setCurrentIndex(index)
-            renderComboBox.activated[str].connect(
-                lambda text=renderComboBox: self.changeRendering(text)
-            )
-            self.renderComboBox = renderComboBox
-            self.grid_layout.addWidget(
-                QLabel('rendering:'), row, self.name_column
-            )
-            self.grid_layout.addWidget(
-                renderComboBox, row, self.property_column
-            )
+        self.layer.events.rendering.connect(self._on_rendering_change)
+        row = self.grid_layout.rowCount()
+        renderComboBox = QComboBox()
+        for render in Rendering:
+            renderComboBox.addItem(str(render))
+        index = renderComboBox.findText(
+            self.layer.rendering, Qt.MatchFixedString
+        )
+        renderComboBox.setCurrentIndex(index)
+        renderComboBox.activated[str].connect(
+            lambda text=renderComboBox: self.changeRendering(text)
+        )
+        self.renderComboBox = renderComboBox
+        self.grid_layout.addWidget(QLabel('rendering:'), row, self.name_column)
+        self.grid_layout.addWidget(renderComboBox, row, self.property_column)
 
         self.setExpanded(False)
-
-    def changeColor(self, text):
-        self.layer.colormap = text
 
     def changeInterpolation(self, text):
         self.layer.interpolation = text
@@ -137,14 +105,6 @@ class QtImageProperties(QtLayerProperties):
                 self.layer.interpolation, Qt.MatchFixedString
             )
             self.interpComboBox.setCurrentIndex(index)
-
-    def _on_colormap_change(self, event):
-        name = self.layer.colormap[0]
-        if name not in self.colormap_combobox._allitems:
-            self.colormap_combobox._allitems.add(name)
-            self.colormap_combobox.addItem(name)
-        if name != self.colormap_combobox.currentText():
-            self.colormap_combobox.setCurrentText(name)
 
     def _on_rendering_change(self, event):
         with self.layer.events.rendering.blocker():
