+++ napari/_qt/layers/util.py
@@ -1,5 +1,6 @@
 from ...layers import Image, Labels, Points, Pyramid, Shapes, Surface, Vectors
 from .qt_base_layer import QtLayerProperties, QtLayerControls
+from .qt_image_base_layer import QtImageBaseProperties
 from .qt_image_layer import QtImageProperties, QtImageControls
 from .qt_points_layer import QtPointsProperties, QtPointsControls
 from .qt_vectors_layer import QtVectorsProperties
@@ -13,7 +14,7 @@ layer_to_properties = {
     Points: QtPointsProperties,
     Pyramid: QtImageProperties,
     Shapes: QtShapesProperties,
-    Surface: QtImageProperties,
+    Surface: QtImageBaseProperties,
     Vectors: QtVectorsProperties,
 }
